+++ ReleaseNotes/release_notes.md
@@ -9,7 +9,9 @@
   1. add `Ostap::MoreRooFit::Hypot` and `var_hypot`
   1. add Fizher's Z-distribution
   1. explicitely delete all created styles, see [here](https://github.com/root-project/root/issues/16918)
-
+  1. add missing methods for `Ostap::Math::Bessel`
+  1. fix occasional overflows in `twosamples.py`
+  
 ## Backward incompatible
     
 ## Bug fixes